# [WIP] Enable LTO: 500 KiB smaller binary, 15% faster build