--- a/script/ci-build-libsass
+++ b/script/ci-build-libsass
@@ -53,15 +53,16 @@ if [ "x$PREFIX" == "x" ]; then
   fi
 fi
 
-# enable address sanitation
-# https://en.wikipedia.org/wiki/AddressSanitizer
-if [ "x$CC" == "xclang" ]; then
-  if [ "x$COVERAGE" != "xyes" ]; then
-    if [ "$TRAVIS_OS_NAME" == "linux" ]; then
-      export EXTRA_CFLAGS="$EXTRA_CFLAGS -fsanitize=address"
-      export EXTRA_CXXFLAGS="$EXTRA_CXXFLAGS -fsanitize=address"
-      export EXTRA_LDFLAGS="$EXTRA_LDFLAGS -fsanitize=address"
-    fi
+# -flto doesn't work on travis due to missing linker plugins.
+export NO_LTO=1
+
+if [[ $CC == clang ]]; then
+  # enable address sanitation
+  # https://en.wikipedia.org/wiki/AddressSanitizer
+  if [[ $COVERAGE != yes && $TRAVIS_OS_NAME == linux ]]; then
+    export EXTRA_CFLAGS="$EXTRA_CFLAGS -fsanitize=address"
+    export EXTRA_CXXFLAGS="$EXTRA_CXXFLAGS -fsanitize=address"
+    export EXTRA_LDFLAGS="$EXTRA_LDFLAGS -fsanitize=address"
   fi
 fi