--- a/script/ci-install-deps
+++ b/script/ci-install-deps
@@ -6,15 +6,4 @@ else
   echo "no dependencies to install"
 fi
 
-if [ "x$AUTOTOOLS" == "xyes" ]; then
-  AUTOTOOLS=yes
-
-  if [ "$TRAVIS_OS_NAME" == "linux" ]; then
-    sudo add-apt-repository -y ppa:rbose-debianizer/automake &> /dev/null
-    sudo apt-get -qq update
-    sudo apt-get -qq install automake
-  fi
-
-fi
-
 exit 0