Metrics:

* `time make -j1 sassc`: 65s -> 56s
* `libsass.a` size: 6.7 MiB ⭢ 24 MiB
* `libsass.so` size: 3.2 MiB ⭢ 2.7 MiB
* sassc binary size: 2.7 MiB ⭢ 2.2 MiB
* Compiling 10 copies of bootstrap 4: 0.45s ⭢ 0.40s

LTO requires a "recent" (2010+) version of `ar` and `ranlib` (to have LTO support by default). Alternatively, one can use `gcc-ar` / `llvm-ar`. In make and automake we try to detect the compiler and set `AR` and `RANLIB`.

LTO can be disabled by setting `NO_LTO=1`.

I still haven't figured out some builds.