@glebm 2018-12-21
I think LTO can work on Travis if we install a newer version of binutils (https://packages.ubuntu.com/xenial/binutils), at least for GCC