@glebm 2018-12-25
Closing, as this is likely to cause build issues for some older distributions.